Moved tags api into own subsection
