+++ python/aristotle-mdr-api/aristotle_mdr_api/v4/urls.py
@@ -1,12 +1,11 @@
-from django.conf.urls import url
+from django.conf.urls import url, include
 from aristotle_mdr_api.v4 import views
 
 urlpatterns = [
+    url(r'tags/', include('aristotle_mdr_api.v4.tags.urls')),
     url(r'issues/$', views.IssueCreateView.as_view(), name='issues_create'),
     url(r'issues/(?P<pk>\d+)/$', views.IssueView.as_view(), name='issues'),
     url(r'issues/(?P<pk>\d+)/updatecomment/$', views.IssueUpdateAndCommentView.as_view(), name='issue_update_and_comment'),
     url(r'issues/comments/$', views.IssueCommentCreateView.as_view(), name='issue_comment'),
     url(r'issues/comments/(?P<pk>\d+)/$', views.IssueCommentRetrieveView.as_view(), name='issue_comment_get'),
-    url(r'tags/(?P<pk>\d+)/$', views.TagView.as_view(), name='tags'),
-    url(r'item/(?P<iid>\d+)/tags$', views.ItemTagUpdateView.as_view(), name='item_tags'),
 ]
